fitness.py works
